Task 2 - finished
Create an HTML gallery page.

At the same time, 1 is displayed on the screen
image and 2 buttons:
Back and forth.
When pressing the buttons, the images should switch in the specified order.

When next / previous
there will be no image, then the corresponding button must be blocked.

Save the image in a previously prepared array.